small refactor to split ch32-data-serde lib.rs
